# server/docs/swagger.yaml
definitions:
...
  cms.MeLetter:
    properties:
      ID:
        description: 主键ID
        type: integer
      applicant:
        description: 申请人
        type: string
      code:
        description: 保单号
        type: string
      court:
        description: 法院
        type: string
      coverage:
        description: 保费
        type: string
      coverageAll:
        description: 保额
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      endCreatedAt:
        description: 结束时间
        type: string
      fileID:
        description: 文件id
        type: string
      info:
        description: 保险信息
        type: string
      policyholder:
        description: 投保人
        type: string
      respondent:
        description: 被申请人
        type: string
      signStatus:
        description: 签署状态
        type: string
      startCreatedAt:
        description: 开始时间
        type: string
      templateFileUrl:
        description: 文件下载
        type: string
      type:
        description: 出单方式
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - applicant
    - court
    - coverage
    - coverageAll
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: cms-gva Swagger API接口文档
  version: v0.0.1
paths:
...
  /meLetter/createMeLetter:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建我的保函
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.MeLetter'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建我的保函
      tags:
      - MeLetter
  /meLetter/deleteMeLetter:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除我的保函
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.MeLetter'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除我的保函
      tags:
      - MeLetter
  /meLetter/deleteMeLetterByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除我的保函
      tags:
      - MeLetter
  /meLetter/findMeLetter:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 申请人
        in: query
        name: applicant
        required: true
        type: string
      - description: 保单号
        in: query
        name: code
        type: string
      - description: 法院
        in: query
        name: court
        required: true
        type: string
      - description: 保费
        in: query
        name: coverage
        required: true
        type: string
      - description: 保额
        in: query
        name: coverageAll
        required: true
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - description: 结束时间
        in: query
        name: endCreatedAt
        type: string
      - description: 文件id
        in: query
        name: fileID
        type: string
      - description: 保险信息
        in: query
        name: info
        type: string
      - description: 投保人
        in: query
        name: policyholder
        type: string
      - description: 被申请人
        in: query
        name: respondent
        type: string
      - description: 签署状态
        in: query
        name: signStatus
        type: string
      - description: 开始时间
        in: query
        name: startCreatedAt
        type: string
      - description: 文件下载
        in: query
        name: templateFileUrl
        type: string
      - description: 出单方式
        in: query
        name: type
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    remeLetter:
                      $ref: '#/definitions/cms.MeLetter'
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询我的保函
      tags:
      - MeLetter
  /meLetter/getMeLetterList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取我的保函列表
      tags:
      - MeLetter
  /meLetter/getMeLetterPublic:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 不需要鉴权的我的保函接口
      tags:
      - MeLetter
  /meLetter/updateMeLetter:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新我的保函
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.MeLetter'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新我的保函
      tags:
      - MeLetter
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"